Easier extraction of all events + example streaming to an analytics DB
This is a breaking changes version!

In this release, the extractor is much more clever about filtering the
events. It is possible to pass an ABI and the exporter will export all
the events that are emitted on the blockchain, which appear in this ABI.

It is possible to filter by topic and address if needed.

There is also an additional ClickHouse service in the example, which
shows how to get the events into a DB for additional analytics.